% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfgattribs.R
\name{cfgattribs}
\alias{cfgattribs}
\title{cfgattribs}
\usage{
cfgattribs(input, output, session, M, D)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{M}{is the meta data connection structure}

\item{D}{is the data connection structure}
}
\description{
Display, edit, create attributes
}
