% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfgattribs.R
\name{cfgattribsUI}
\alias{cfgattribsUI}
\title{cfgattribsUI}
\usage{
cfgattribsUI(id, M)
}
\arguments{
\item{id}{is caller id}

\item{M}{is the meta data connection structure}
}
\description{
UI for Display, edit, create attributes
}
