% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfgmetrics.R
\name{cfgmetricsUI}
\alias{cfgmetricsUI}
\title{cfgmetricsUI}
\usage{
cfgmetricsUI(id, M)
}
\arguments{
\item{id}{is caller id}

\item{M}{is the meta data connection structure}
}
\description{
UI for Display, edit, create measures
}
