% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmodel.R
\name{dmodel}
\alias{dmodel}
\title{dmodel}
\usage{
dmodel(input, output, session, M, D)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{M}{is the meta data connection structure}

\item{D}{is the data connection structure}
}
\description{
Display, edit, create data model / table joins
}
