% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmodel.R
\name{dmodelUI}
\alias{dmodelUI}
\title{cfgattribsUI}
\usage{
dmodelUI(id, M)
}
\arguments{
\item{id}{is caller id}

\item{M}{is the meta data connection structure}
}
\description{
UI for Display, edit, create attributes
}
